fix(css): add IE-specific styling to address itin rendering
Apparently Internet Explorer handles table height rendering differently than other browsers. Because
the itinerary leg line uses table-cell display, we need some IE-specific way to handle the height
issue. This is a bit hacky, but uses a media query for IE to set the height to a very large value
(with overflow hidden) to force a more or less correct display of the leg.

refs #69